feat: partially implement venues

+++ src/routes/api/venues/floors/[id].ts
@@ -3,18 +3,6 @@ import type { FloorLevel, Locals } from '$lib/types';
 import { mapbox } from '$lib/variables';
 import type { RequestHandler } from '@sveltejs/kit';
 
-// get floor
-export const get: RequestHandler<Locals> = async ({ params }) => {
-  const client = await clientPromise;
-  const collection = client.db().collection<FloorLevel>('floors');
-  const floor = await collection.findOne<FloorLevel>({ id: params.id }, { projection: { _id: 0 } });
-
-  return {
-    status: 200,
-    body: JSON.stringify(floor),
-  };
-};
-
 // put floor
 export const put: RequestHandler<Locals, string> = async ({ params, body }) => {
   const floor: FloorLevel = JSON.parse(body);
@@ -27,7 +15,7 @@ export const put: RequestHandler<Locals, string> = async ({ params, body }) => {
 
   return {
     status: 200,
-    body: JSON.stringify(updatedFloor),
+    body: updatedFloor as any,
   };
 };
 
@@ -37,6 +25,15 @@ export const del: RequestHandler<Locals> = async ({ params }) => {
   const collection = client.db().collection<FloorLevel>('floors');
 
   const floor = await collection.findOne({ id: params.id });
+
+  if (!floor) {
+    return {
+      status: 404,
+      body: {
+        error: 'floor not found',
+      },
+    };
+  }
 
   const response = await fetch(`${mapbox.baseTilesetUrl}/${floor.tileset}?access_token=${mapbox.uploadToken}`, { method: 'DELETE' });
   const result = await response.json();
